Move files and rename packages in scala 2.
