+++ src/main/scala-2/scalats/package.scala
@@ -1,5 +1,3 @@
-package com.mpc
-
 package object scalats {
   private[scalats] def decapitalize(s: String): String = s.take(1).toLowerCase ++ s.drop(1)
   private[scalats] def unionOrdName(s: String): String = decapitalize(s) ++ "Ord"
